[iris] Move FakeStatsTable to iris.test_util per shared-fake policy
TESTING.md places shared fakes in src/iris/test_util.py (or a backend fake
module), not in a conftest other suites import from.
