+++ lib/iris/tests/cluster/backends/k8s/test_task_event.py
@@ -25,8 +25,7 @@ from iris.cluster.backends.k8s.tasks import (
 from iris.cluster.controller.task_state import RunningTaskEntry
 from iris.cluster.platforms.k8s.types import K8sResource
 from iris.cluster.types import JobName
-
-from tests.cluster.conftest import FakeStatsTable
+from iris.test_util import FakeStatsTable
 
 from .conftest import (
     KUEUE_UNADMITTED_MSG,
